MultiScaleOT.AbstractMeasure — Type

AbstractMeasure{D}

Type from which GridMeasure{D}, CloudMeasure{D} inherit. It is left unexported to avoid potential conflicts with MeasureTheory.jl, which has another type with the same name.

MultiScaleOT.CloudMeasure — Type

CloudMeasure{D} <: AbstractMeasure{D}

A measure supported on an unstructured D-dimensional point cloud. Its attributes are:

points::Matrix{Float64}: points[:,j] are the coordinates of point j.

weights::Vector{Float64}: weights[j] is the mass of point j.

extent::NTuple{D, Tuple{Float64, Float64}}: extent[i] are lower and upper bounds on the i-th row of points.

MultiScaleOT.GridMeasure — Type

GridMeasure{D} <: AbstractMeasure{D}

A measure supported on a D-dimensional grid. Its attributes are:

points::Matrix{Float64}: points[:,j] are the coordinates of point j.

weights::Vector{Float64}: weights[j] is the mass of point j.

gridshape::NTuple{D,Int}: shape of the grid, i.e. (length of x1, length of x2,...)

MultiScaleOT.MultiScaleMeasure — Type

MultiScaleMeasure{<:AbstractMeasure}

Type encoding a measure at different resolutions, and how to refine a coarse version to a finer version. Inner representation is

mutable struct MultiScaleMeasure{M<:AbstractMeasure{D} where D} 
    depth::Int # Number of resolutions
    measures::Vector{M} # `measures[i]` gives resolution `i`
    refinements::Vector{Vector{Vector{Int}}} # `refinements[i][j]` yields the indices of
                                             # `mesures[i+1]` into which the atom
                                             # `j` of `measures[i]` is splitted
end

MultiScaleOT.MultiScaleMeasure — Method

MultiScaleMeasure(mu[; depth])

Compute the MultiScaleMeasure with depth levels and finest resolution given by mu.

MultiScaleOT.KL — Method

KL(a, b)

Computes KL divergence between vectors a and b. Yield Inf if a[i]>0, b[i]=0 for some i.

MultiScaleOT.PD_gap_dense — Method

PD_gap_dense(a, b, P, c, X, Y, μ, ν, ε)

Compute primal_score - dual_score using the dense versions of these functions.

MultiScaleOT.PD_gap_sparse — Method

PD_gap_dense(a, b, P, c, X, Y, μ, ν, ε)

Compute primal_score - dual_score using the dense versions of these functions.

MultiScaleOT.compute_multiscale_depth — Method

compute_multiscale_depth(mu::GridMeasure{D})

Compute the minimum multiscale depth to refine a measure consisting of a single point to mu by dividing each point iteratively into (at most) 2^D points

MultiScaleOT.constant_schedule — Method

constant_schedule(value, N)

Repeat value N times

MultiScaleOT.dual_score_dense — Method

dual_score_dense(a, b, c, X, Y, μ, ν, ε)

Compute the dual score of (a,b) for the entropic problem with cost function c and regularization ε, using all points in the product space.

MultiScaleOT.dual_score_sparse — Method

dual_score_sparse(a, b, c, X, Y, μ, ν, ε, P::SparseMatrixCSC)

Compute the dual score of (a,b) for the entropic problem with cost function c and regularization ε, using only the indices (i,j) where P is non-zero.

MultiScaleOT.dual_score_sparse_eps0 — Method

dual_score_sparse_eps0(P, c, X, Y, μ, ν, ε, P; threshold = 1e-8)

Unregularized dual cost. Yields infinite if some entry of the form a[i]+b[j]-C[i,j] in the support of P is bigger than threshold.

MultiScaleOT.euclidean_barycenter — Method

euclidean_barycenter(X, w)

Compute the euclidean barycenter of the columns of X with weights w

MultiScaleOT.fine_to_coarse — Method

fine_to_coarse(mu::GridMeasure{D}, cellsize)

Compute a coarse approximation of mu by combining cellsize × ... × cellsize points together. Return the coarser Measure new_mu and the partition cells, such that sum(mu.weights[cells[i]]) == new_mu.weights[i].

MultiScaleOT.flat_grid — Method

flat_grid(xs...)

Create a grid with the arguments xs... and flatten its points to a matrix. Each column of the matrix represents one point on the grid, and the first coordinate runs faster.

Examples

julia> x = [1, 2]; y = [2, 3];
julia> flat_grid(x, y)

2×4 Matrix{Int64}:
 1  2  1  2
 2  2  3  3

MultiScaleOT.get_cells — Method

get_cells(shape, cellsize, offset=0)

Call either get_cells_1D, get_cells_2D or get_cells_3D depending on the size of shape. Higher dimensional not implemented.

MultiScaleOT.get_cells_1D — Function

get_cells_1D(N, cellsize, offset)

Divide [1,...,N] into cells of size cellsize, with the first cell having offset indices. Return the list of cells and its shape.

Examples

julia> get_cells_1D(4, 2, 0)
([[1, 2], [3, 4]], (2,))

julia> get_cells_1D(4, 2, 1)
([[1], [2, 3], [4]], (3,))

MultiScaleOT.get_cells_2D — Method

get_cells_2D(shape, cellsize, offset)

Divide [1,...,shape[1]]×[1,...,shape[2]] into cells of size (cellsize, cellsize), with the given offset. The order of the basic cells is lexicographic, with the first index running faster. Return the list of cells and its shape.

Examples

julia> get_cells_2D((4,4), 2, 0)
([[1, 2, 5, 6], [3, 4, 7, 8], [9, 10, 13, 14], [11, 12, 15, 16]], (2, 2))

MultiScaleOT.get_cells_3D — Method

get_cells_3D(shape, cellsize, offset)

Divide [1,...,shape[1]]×[1,...,shape[2]]×[1,...,shape[3]] into cells of size (cellsize, cellsize, cellsize), with the given offset. The order of the basic cells is lexicographic, with the first index running faster. Return the list of cells and its shape.

MultiScaleOT.get_grid_nodes — Method

get_grid_nodes(X, shapeX)

For X of size (N, prod(shapeX)) on a grid, get its 1-dimensional factors.

Examples

julia> X = [1 2 1 2 1 2
            1 1 2 2 3 3];
julia> get_grid_nodes(X, (2, 3))
([1, 2], [1, 2, 3])

MultiScaleOT.get_kernel! — Method

get_kernel!(C, a, b, μ, ν, ε)

Compute inplace the Gibbs energy of matrix C, current duals a and b and scale it with marginals μ and ν.

MultiScaleOT.get_kernel — Method

get_kernel!(K, μ, ν, ε)

Compute the Gibbs energy of matrix C, current duals a and b and scale it with marginals μ and ν.

MultiScaleOT.get_missing_dual_potential — Method

get_missing_dual_potential(C, ϕ, ρ, ε, dim)

Compute the conjugate potential of ϕ with regularization ε with marginal ρ. dim is the numeral of this potential.

Examples

If (α, β) are the dual potentials for sinkhorn(C, μ, ν, ε), then α == get_missing_dual_potential(C, β, ν, ε, 1) and β == get_missing_dual_potential(C, α, μ, ε, 2).

MultiScaleOT.get_product_list — Method

get_product_list(it1, it2, it3, N1, N2)

Obtain the list of flattened indices of the cartesian indices (i,j,k), for k in it3, j in it2 and i in it1, assuming that the total dimension of the first and second axes are respectively N1 and N2.

MultiScaleOT.get_product_list — Method

get_product_list(it1, it2, N1)

Obtain the list of flattened indices of the cartesian indices (i,j), for j in it2 and i in it1, assuming that the total dimension of the first axis is N1.

MultiScaleOT.get_stabilized_kernel — Method

get_stabilized_kernel(c::Function, a, b, ε, X, Y, μ, ν, θ[, colptr0, roval0])

Compute a truncated, stabilized kernel, following Section 3.3 of https://arxiv.org/abs/1610.06519, using the dual variables (a, b), the threshold θ and a previous estimate on the neighborhood given by (colptr0, rowval0). These must be interpreted as the representation of a CSC sparse matrix, on a subset of whose entries the truncated kernel will lie.

If colptr0, rowval0 are not given, it is assumed that the all entries must be inspected.

MultiScaleOT.isthere_nan_or_inf — Method

isthere_nan_or_inf(v)

Check if there is NaN or Inf in iterable v

MultiScaleOT.l1 — Method

l1(a, b)

Compute ℓ^1 distance between a and b. Does not allocate. If the sizes of the vectors are known with certainty, l1_inbounds might be preferred.

MultiScaleOT.l1_inbounds — Method

l1_inbounds(a, b)

Compute ℓ^1 distance between a and b. Does not allocate nor perform bound checks on the vectors

MultiScaleOT.log_sinkhorn! — Method

log_sinkhorn!(a, b, μ, ν, C, ε; kwargs...)

Run a log-stabilized implementation of the Sinkhorn algorithm on potentials a and b (first iteration updates a), with marginals mu and nu, cost matrix C and regularization strength ε, until some maximum error or number of iterations is achieved.

Optional arguments are:

max_iter: Maximum number of iterations.

max_error: Desired error.

max_error_rel: Whether the max_error is relative to the mass of μ.

verbose: Whether convergence updates are desired.

MultiScaleOT.log_sinkhorn! — Method

log_sinkhorn!(a, b, μ, ν, Ka, Kb, K, Niter)

Run a log-stabilized implementation of the Sinkhorn algorithm for Niter iterations. Mutate its arguments.

MultiScaleOT.logsumexp! — Method

logsumexp!(y, A, b, ε, dims)

Compute the softmin of A - b with regularization ε along the dims dimension, where b is subtracted also along the dims dimension. Modify the array y.

MultiScaleOT.logsumexp — Method

logsumexp(x, b, ε)

Compute the softmin of x-b with regularization ε.

MultiScaleOT.logsumexp — Method

logsumexp(x, ε)

Compute the softmin of x with regularization ε.

MultiScaleOT.logsumexp — Method

logsumexp(A, b, ε, dims)

Compute the softmin of A - b with regularization ε along the dims dimension, where b is subtracted also along the dims dimension.

MultiScaleOT.logsumexp — Method

logsumexp(A, ε, dims)

Compute the softmin of A with regularization ε along the dims dimension.

MultiScaleOT.lp — Method

lp(a, b, p)

Compute the ℓ^p distance between a and b. Does not allocate. If the sizes of the vectors are known with certainty, lp_inbounds might be preferred.

MultiScaleOT.lp_inbounds — Method

lp_inbounds(a, b, p)

Compute the ℓ^p distance between a and b. Does not allocate nor perform bounds checks on the vectors

MultiScaleOT.lpp — Method

lpp(a, b, p)

Compute the p-th power of the ℓ^p distance between a and b. Does not allocate. If the sizes of the vectors are known with certainty, lpp_inbounds might be preferred.

MultiScaleOT.lpp_inbounds — Method

lpp_inbounds(a, b, p)

Compute p-th power of the ℓ^p distance between a and b. Does not allocate nor perform bounds checks on the vectors

MultiScaleOT.make_schedule — Method

make_schedule(; nt...)

Generate a StructArray` representing a schedule, this is, columns represent the values a parameter, is taking on each step of the algorithm; conversely, rows represent the set of parameters to be used at a given stage of the algorithm.

The elements in nt are arrays of singletons. In the later case they are repeated to match the length of the array ones.

Examples

julia> sinkhornschedule = makeschedule( ε = [4, 2, 2, 1], N = [64, 64, 32, 32], θ = 1e-20, max_error = 1e-6 )

4-element StructArray(::Vector{Int64}, ::Vector{Int64}, ::Vector{Float64}, ::Vector{Float64}) with eltype NamedTuple{(:ε, :N, :θ, :maxerror), Tuple{Int64, Int64, Float64, Float64}}: (ε = 4, N = 64, θ = 1.0e-20, maxerror = 1.0e-6) (ε = 2, N = 64, θ = 1.0e-20, maxerror = 1.0e-6) (ε = 2, N = 32, θ = 1.0e-20, maxerror = 1.0e-6) (ε = 1, N = 32, θ = 1.0e-20, max_error = 1.0e-6)

MultiScaleOT.mean — Method

mean(x, J)

Mean of x[J], non-allocating

MultiScaleOT.normalize! — Function

normalize!(a[, mass = 1])

Normalize the vector a so that it adds up to mass. Does not perform any check on the positivity of a.

MultiScaleOT.pad_extrapolate — Method

pad_extrapolate(x::AbstractVector{T})

Extrapolate a vector of length N at both ends. The result is a vector of length N+2.

Examples

julia> pad_extrapolate([1,2,3])
5-element Vector{Int64}:
 0
 1
 2
 3
 4

MultiScaleOT.pad_extrapolate — Method

pad_extrapolate(A::Matrix)

Extrapolate a (N,M) matrix along its boundary. The result is a (N+2, M+2) matrix.

Examples

julia> pad_extrapolate([1 2; 3 4])
4×4 Matrix{Float64}:
 -2.0  -1.0  0.0  1.0
  0.0   1.0  2.0  3.0
  2.0   3.0  4.0  5.0
  4.0   5.0  6.0  7.0

MultiScaleOT.primal_score_dense — Method

primal_score_dense(P, c, X, Y, μ, ν, ε)

Compute the primal score of P for the entropic problem with cost function c and regularization ε, using all entries of P. Works for P both dense and sparse.

MultiScaleOT.primal_score_sparse — Method

primal_score_sparse(P::SparseMatrixCSC, c, X, Y, μ, ν, ε)

Compute the primal score of P for the entropic problem with cost function c and regularization ε, using only the stored entries of P.

MultiScaleOT.primal_score_sparse_eps0 — Method

primal_score_sparse_eps0(P, c, X, Y, μ, ν, ε)

Unregularized primal cost. To be called by primal_score_sparse

MultiScaleOT.refine_support — Method

refine_support(colptr0, rowval0, refinementX, refinementY)

Seeing colptr0, rowval0 as the support of a sparse matrix K0, obtain a new support colptr, rowval that is the support of the matrix K resulting from the refinements refinementX, refinementY

MultiScaleOT.scaling_schedule — Method

scaling_schedule(depth, target_value, Nsteps, factor; last_iter = [])

Generate the vector

target_value.* [factor^(depth*(Nsteps-1)), 
                factor^(depth*(Nsteps-1)-1),
                factor^(depth*(Nsteps-1)-2),
                ...
                factor^((depth-1)*(Nsteps-1)),
                factor^((depth-1)*(Nsteps-1)),
                factor^((depth-1)*(Nsteps-1)-1),
                ...
                factor^((depth-2)*(Nsteps-1)),
                factor^((depth-2)*(Nsteps-1)),
                ...
                ...
                1
                ;
                last_iter]

Useful for generating scaling schedules for the regularization ε.

Examples

julia> eps_target = 1.; factor = 2.; Nsteps = 3;
julia> eps_schedule = scaling_schedule(depth, eps_target, Nsteps, factor)
9-element Vector{Float64}:
 64.0
 32.0
 16.0
 16.0
  8.0
  4.0
  4.0
  2.0
  1.0

MultiScaleOT.sinkhorn! — Method

sinkhorn!(u, v, μ, ν, K, KT, Niter)

Run Niter iterations of the Sinkhorn algorithm inplace on the scaling factors (u,v), with marginals μ and ν and Gibbs entropy given by K and its transpose KT. At the end of the iterates the coupling u.*K.*v'has exact first marginal, i.e.,sum(u .* K .* v', dims = 2)[:] == μ`.

MultiScaleOT.sinkhorn_stabilized! — Method

sinkhorn_stabilized!(a, b, μ, ν, K, ε; kwargs...)

Implement the stabilized Sinkhorn algorithm of https://arxiv.org/abs/1610.06519.

MultiScaleOT.template_schedule — Method

template_schedule(depth, template, factors; last_iter = [])

Generate the vector [factors[1].*template; ... ; factors[depth].*template; last_iter] Useful to define a schedule with a certain substructure. When factors is a singleton, it is transformed into fill(factors, depth). When template is an Int, it is transformed into ones(template).

Examples

julia> depth = 3; Ns = [1, 2, 4]; Nsteps = 2;
julia> template_schedule(depth, Nsteps, Ns)
6-element Vector{Int64}:
 1
 1
 2
 2
 4
 4

julia> depth = 3; maxiters_template = [2000, 1000]; factor = 1;
julia> template_schedule(depth, maxiters_template, factor)
6-element Vector{Int64}:
 2000
 1000
 2000
 1000
 2000
 1000

MultiScaleOT.update_current_Y_marginal! — Method

update_current_Y_marginal!(νK, v, KT, u, ν)

Update the current Y marginal of the coupling and return the error with respect to the true marginal ν.

MultiScaleOT.DEFAULT_PARAMETERS — Constant

DEFAULT_PARAMETERS

Default parameters for the multiscale routines